Start refactor of houses UI, move magic guilds to separate location
