+++ src/GUI/GUIWindow.cpp
@@ -319,215 +319,6 @@ void GUIWindow::DrawMessageBox(bool inside_game_viewport) {
     }
 }
 
-void GUIWindow::HouseDialogManager() {
-    if (!window_SpeakInHouse) {
-        return;
-    }
-    GUIWindow pWindow = *this;
-    pWindow.uFrameWidth -= 18;
-    pWindow.uFrameZ -= 18;
-    render->DrawTextureNew(477 / 640.0f, 0, game_ui_dialogue_background);
-    render->DrawTextureNew(468 / 640.0f, 0, game_ui_right_panel_frame);
-
-    if (pDialogueNPCCount != uNumDialogueNPCPortraits || !uHouse_ExitPic) {
-        const char *pHouseName = buildingTable[window_SpeakInHouse->wData.val - 1].pName;
-        if (pHouseName) {
-            if (current_screen_type != CURRENT_SCREEN::SCREEN_SHOP_INVENTORY) {
-                int v3 = 2 * pFontCreate->GetHeight() - 6 -
-                    pFontCreate->CalcTextHeight(pHouseName, 130, 0);
-                if (v3 < 0) v3 = 0;
-                pWindow.DrawTitleText(pFontCreate, 0x1EAu, v3 / 2 + 4, colorTable.White.c16(),
-                    buildingTable[window_SpeakInHouse->wData.val - 1].pName, 3);
-            }
-        }
-    }
-
-    pWindow.uFrameWidth += 8;
-    pWindow.uFrameZ += 8;
-    if (!pDialogueNPCCount) {
-        if (in_current_building_type == BuildingType_Jail) {
-            JailDialog();
-            if (pDialogueNPCCount == uNumDialogueNPCPortraits &&
-                uHouse_ExitPic) {
-                render->DrawTextureNew(556 / 640.0f, 451 / 480.0f,
-                    dialogue_ui_x_x_u);
-                render->DrawTextureNew(476 / 640.0f, 451 / 480.0f,
-                    dialogue_ui_x_ok_u);
-            } else {
-                render->DrawTextureNew(471 / 640.0f, 445 / 480.0f,
-                    ui_exit_cancel_button_background);
-            }
-            return;
-        }
-        if (!current_npc_text.empty()) {
-            GUIWindow pDialogWindow;
-            pDialogWindow.uFrameWidth = 458;
-            pDialogWindow.uFrameZ = 457;
-            int pTextHeight = pFontArrus->CalcTextHeight(
-                current_npc_text, pDialogWindow.uFrameWidth, 13);
-            int v6 = pTextHeight + 7;
-            render->DrawTextureCustomHeight(8 / 640.0f,
-                (352 - (pTextHeight + 7)) / 480.0f,
-                ui_leather_mm7, pTextHeight + 7);
-            render->DrawTextureNew(8 / 640.0f, (347 - v6) / 480.0f, _591428_endcap);
-            window_SpeakInHouse->DrawText(pFontArrus, {13, 354 - v6}, 0, pFontArrus->FitTextInAWindow(current_npc_text, pDialogWindow.uFrameWidth, 13), 0, 0, 0);
-        }
-        if (uNumDialogueNPCPortraits <= 0) {
-            if (pDialogueNPCCount == uNumDialogueNPCPortraits &&
-                uHouse_ExitPic) {
-                render->DrawTextureNew(556 / 640.0f, 451 / 480.0f,
-                    dialogue_ui_x_x_u);
-                render->DrawTextureNew(476 / 640.0f, 451 / 480.0f,
-                    dialogue_ui_x_ok_u);
-            } else {
-                render->DrawTextureNew(471 / 640.0f, 445 / 480.0f,
-                    ui_exit_cancel_button_background);
-            }
-            return;
-        }
-
-        unsigned int v10 = 0;
-
-        for (int v8 = 0; v8 < uNumDialogueNPCPortraits; ++v8) {
-            render->DrawTextureNew(
-                (pNPCPortraits_x[uNumDialogueNPCPortraits - 1][v8] - 4) /
-                640.0f,
-                (pNPCPortraits_y[uNumDialogueNPCPortraits - 1][v8] - 4) /
-                480.0f,
-                game_ui_evtnpc);
-            render->DrawTextureNew(
-                pNPCPortraits_x[uNumDialogueNPCPortraits - 1][v8] / 640.0f,
-                pNPCPortraits_y[uNumDialogueNPCPortraits - 1][v8] / 480.0f,
-                pDialogueNPCPortraits[v8]);
-            if (uNumDialogueNPCPortraits < 4) {
-                std::string pTitleText;
-                int v9 = 0;
-                if (v8 + 1 == uNumDialogueNPCPortraits && uHouse_ExitPic) {
-                    pTitleText = pMapStats->pInfos[uHouse_ExitPic].pName;
-                    v9 = 94 * v8 + SIDE_TEXT_BOX_POS_Y;
-                } else {
-                    if (!v8 && dword_591080) {
-                        pTitleText = (char*)buildingTable[window_SpeakInHouse->wData.val - 1].pProprieterTitle;
-                        pWindow.DrawTitleText(pFontCreate, SIDE_TEXT_BOX_POS_X, SIDE_TEXT_BOX_POS_Y, colorTable.EasternBlue.c16(), pTitleText, 3);
-                        continue;
-                    }
-                    pTitleText = HouseNPCData[v8 + 1 - (dword_591080 != 0)]->pName;
-                    v9 = pNPCPortraits_y[uNumDialogueNPCPortraits - 1][v8] + pDialogueNPCPortraits[v8]->GetHeight() + 2;
-                }
-                v10 = v9;
-                pWindow.DrawTitleText(pFontCreate, SIDE_TEXT_BOX_POS_X, v10, colorTable.EasternBlue.c16(), pTitleText, 3);
-            }
-        }
-        if (pDialogueNPCCount == uNumDialogueNPCPortraits && uHouse_ExitPic) {
-            render->DrawTextureNew(556 / 640.0f, 451 / 480.0f,
-                dialogue_ui_x_x_u);
-            render->DrawTextureNew(476 / 640.0f, 451 / 480.0f,
-                dialogue_ui_x_ok_u);
-        } else {
-            render->DrawTextureNew(471 / 640.0f, 445 / 480.0f,
-                ui_exit_cancel_button_background);
-        }
-        return;
-    }
-
-    int v4 = pDialogueNPCCount - 1;
-    render->DrawTextureNew((pNPCPortraits_x[0][0] - 4) / 640.0f,
-        (pNPCPortraits_y[0][0] - 4) / 480.0f,
-        game_ui_evtnpc);
-    render->DrawTextureNew(pNPCPortraits_x[0][0] / 640.0f,
-        pNPCPortraits_y[0][0] / 480.0f,
-        pDialogueNPCPortraits[v4]);
-    if (current_screen_type == CURRENT_SCREEN::SCREEN_SHOP_INVENTORY) {
-        CharacterUI_InventoryTab_Draw(&pParty->activeCharacter(), true);
-        if (pDialogueNPCCount == uNumDialogueNPCPortraits && uHouse_ExitPic) {
-            render->DrawTextureNew(556 / 640.0f, 451 / 480.0f,
-                dialogue_ui_x_x_u);
-            render->DrawTextureNew(476 / 640.0f, 451 / 480.0f,
-                dialogue_ui_x_ok_u);
-        } else {
-            render->DrawTextureNew(471 / 640.0f, 445 / 480.0f,
-                ui_exit_cancel_button_background);
-        }
-        return;
-    }
-    if (v4 || !dword_591080) {  // emerald isle ship before quest's done   / на
-                                // изумрудном острове заходит на корабле пока не
-                                // выполнены квесты
-        SimpleHouseDialog();
-    } else {
-        pWindow.DrawTitleText(
-            pFontCreate, SIDE_TEXT_BOX_POS_X, SIDE_TEXT_BOX_POS_Y, colorTable.EasternBlue.c16(), NameAndTitle(
-                buildingTable[window_SpeakInHouse->wData.val - 1].pProprieterName,
-                buildingTable[window_SpeakInHouse->wData.val - 1].pProprieterTitle
-            ), 3);
-        switch (in_current_building_type) {
-        case BuildingType_WeaponShop:
-            WeaponShopDialog();
-            break;
-        case BuildingType_ArmorShop:
-            ArmorShopDialog();
-            break;
-        case BuildingType_MagicShop:
-            MagicShopDialog();
-            break;
-        case BuildingType_AlchemistShop:
-            AlchemistDialog();
-            break;
-        case BuildingType_FireGuild:
-        case BuildingType_AirGuild:
-        case BuildingType_WaterGuild:
-        case BuildingType_EarthGuild:
-        case BuildingType_SpiritGuild:
-        case BuildingType_MindGuild:
-        case BuildingType_BodyGuild:
-        case BuildingType_LightGuild:
-        case BuildingType_DarkGuild:
-        case BuildingType_MirroredPath:
-            GuildDialog();
-            break;
-        case BuildingType_MercenaryGuild:
-            MercenaryGuildDialog();
-            break;
-        case BuildingType_TownHall:
-            TownHallDialog();
-            break;
-        case BuildingType_Tavern:
-            TavernDialog();
-            break;
-        case BuildingType_Bank:
-            BankDialog();
-            break;
-        case BuildingType_Temple:
-            TempleDialog();
-            break;
-        case BuildingType_Stables:
-        case BuildingType_Boats:
-            TravelByTransport();
-            break;
-        case BuildingType_Training:
-            // __debugbreak(); // param was passed via pTmpBuf, investiage
-            // ?? no idea why this could pass an argument - its always reset
-            TrainingDialog("");
-            break;
-        case BuildingType_Jail:
-            JailDialog();
-            break;
-        default:
-            // __debugbreak();//New BuildingType (if enter Boat)
-            break;
-        }
-    }
-    if (pDialogueNPCCount == uNumDialogueNPCPortraits && uHouse_ExitPic) {
-        render->DrawTextureNew(556 / 640.0f, 451 / 480.0f,
-            dialogue_ui_x_x_u);
-        render->DrawTextureNew(476 / 640.0f, 451 / 480.0f,
-            dialogue_ui_x_ok_u);
-    } else {
-        render->DrawTextureNew(471 / 640.0f, 445 / 480.0f,
-            ui_exit_cancel_button_background);
-    }
-}
-
 std::string MakeDateTimeString(GameTime time) {
     int seconds = time.GetSecondsFraction();
     int minutes = time.GetMinutesFraction();
