--- src/GUI/UI/UIGuilds.h
@@ -1,6 +1,0 @@
-#pragma once
-
-/**
- * @offset 0x4BC8D5
- */
-void generateSpellBooksForGuild();
